requirements.txt
fastapi==0.115.8
pythainlp==5.0.5
langchain-unstructured==0.1.6
langchain-huggingface==0.1.2
langchain-community==0.3.16
langchain==0.3.17
langchain-core==0.3.33
langchain-google-genai==2.0.9
langgraph==0.2.69
langgraph-checkpoint-sqlite==2.0.3
python-dotenv==1.0.1
unstructured==0.16.17
faiss-cpu==1.10.0
uvicorn==0.34.0